Unified MCP23017 with GPIO wrapper
